+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## edge
 * remove Rails dependencies by using rinku & tag_helper
+* alt for images
 
 ## 1.3.7, released 2011-12-04
 * require activerecord only if available