+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## edge
 * remove Rails dependencies by using rinku & tag_helper
-* alt for images
+* image filter supports attributes
+* html_safe string
+
 
 ## 1.3.7, released 2011-12-04
 * require activerecord only if available