+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.1, released 2011-07-09
+* AutoHtml gets module method: AutoHtml.auto_html(...)
+* AutoHtmlFor producing html_safe
+* on blank filter list return input instead of nil
+
 ## 1.4.0, released 2011-07-09
 * remove Rails dependencies by using rinku & tag_helper
 * image filter supports attributes